tests/wpt/tests/css/css-anchor-position/anchor-name-multicol-001.html

<!DOCTYPE html>
<title>Anchors in a different containing block in multicol</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#propdef-anchor-name">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#anchor-size">
<link rel="author" href="mailto:kojii@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/test-common.js"></script>
<style>
.relpos {
  position: relative;
}
.columns {
  column-count: 2;
  column-fill: auto;
  column-gap: 10px;
  column-width: 100px;
  width: 210px;
  height: 100px;
}
.spacer {
  height: 10px;
  background: pink;
}
.anchor1 {
  anchor-name: --a1;
  margin-left: 10px;
  width: 40px;
  height: 20px;
  background: orange;
}
.target {
  position: absolute;
  left: anchor(--a1 left);
  top: anchor(--a1 top);
  width: anchor-size(--a1 width);
  height: anchor-size(--a1 height);
  background: lime;
  opacity: .3;
}
</style>
<body onload="checkLayoutForAnchorPos('.target')">
  <div class="columns">
    <div class="relpos">
      <div class="spacer"></div>
      <div class="anchor1"></div>
    </div>
    <div class="relpos">
      <div class="target"
           data-offset-x=0 data-offset-y=0
           data-expected-width=0 data-expected-height=0></div>
    </div>
  </div>
</body>
